+++ packages/apollo-cache-inmemory/src/inMemoryCache.ts
@@ -41,6 +41,10 @@ export class InMemoryCache extends ApolloCache<NormalizedCache> {
   private watches: Cache.WatchOptions[] = [];
   private addTypename: boolean;
 
+  // Set this while in a transaction to prevent broadcasts...
+  // don't forget to turn it back on!
+  private silenceBroadcast: boolean = false;
+
   constructor(config: ApolloReducerConfig = {}) {
     super();
     this.config = { ...defaultConfig, ...config };
@@ -141,18 +145,32 @@ export class InMemoryCache extends ApolloCache<NormalizedCache> {
 
   public performTransaction(transaction: Transaction<NormalizedCache>) {
     // TODO: does this need to be different, or is this okay for an in-memory cache?
+
+    let alreadySilenced = this.silenceBroadcast;
+    this.silenceBroadcast = true;
+
     transaction(this);
+
+    if (!alreadySilenced) {
+      // Don't un-silence since this is a nested transaction
+      // (for example, a transaction inside an optimistic record)
+      this.silenceBroadcast = false;
+    }
+
+    this.broadcastWatches();
   }
 
   public recordOptimisticTransaction(
     transaction: Transaction<NormalizedCache>,
     id: string,
   ) {
+    this.silenceBroadcast = true;
+
     const before = this.extract(true);
 
     const orig = this.data;
     this.data = { ...before };
-    transaction(this);
+    this.performTransaction(transaction);
     const after = this.data;
     this.data = orig;
 
@@ -169,6 +187,8 @@ export class InMemoryCache extends ApolloCache<NormalizedCache> {
       transaction,
       data: patch,
     });
+
+    this.silenceBroadcast = false;
 
     this.broadcastWatches();
   }
@@ -224,12 +244,18 @@ export class InMemoryCache extends ApolloCache<NormalizedCache> {
   }
 
   private broadcastWatches() {
+    // Skip this when silenced (like inside a transaction)
+    if (this.silenceBroadcast) return;
+
     // right now, we invalidate all queries whenever anything changes
-    this.watches.forEach(c => {
+    this.watches.forEach((c: Cache.WatchOptions) => {
       const newData = this.diff({
         query: c.query,
         variables: c.variables,
-        previousResult: c.previousResult(),
+
+        // TODO: previousResult isn't in the types - this will only work
+        // with ObservableQuery which is in a different package
+        previousResult: (c as any).previousResult && c.previousResult(),
         optimistic: c.optimistic,
       });